// tile-detect-api/lib/types.ts
export type Suit = "man" | "pin" | "sou";
export type SuitedValue = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 | 9;
export type WindValue = "east" | "south" | "west" | "north";
export type DragonValue = "haku" | "hatsu" | "chun";
export type HonorValue = WindValue | DragonValue;

export interface SuitedTile {
  suit: Suit;
  value: SuitedValue;
  isAka?: boolean; // aka dora (red 5)
}

export interface HonorTile {
  suit: "honor";
  value: HonorValue;
}

export type Tile = SuitedTile | HonorTile;